template_sweeps/requirements.txt
PyYAML
torch
wandb
numpy
scikit-learn
